Update:list to app demo

+++ lib/widgit/card_body_widget.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:thethao12/Modal/items.dart';
+import 'package:confirm_dialog/confirm_dialog.dart';
 
 class CardBody extends StatelessWidget {
   final DataItems item;
@@ -16,7 +17,7 @@ class CardBody extends StatelessWidget {
       width: 350,
       height: 70,
       decoration: BoxDecoration(
-        color: const Color(0xffDFDFDF),
+        color: item.color,
         borderRadius: BorderRadius.circular(12),
       ),
       child: Padding(
@@ -26,7 +27,6 @@ class CardBody extends StatelessWidget {
           children: [
             Text(
               item.name,
-
               style: TextStyle(
                 fontSize: 20,
                 color: Color.fromARGB(255, 0, 0, 0),
@@ -35,9 +35,21 @@ class CardBody extends StatelessWidget {
               ),
             ),
             InkWell(
-              onTap: () {
-                handleDeleteTask(item.id);
+              onTap: () async {
+                if (await confirm(
+                  context,
+                  title: const Text('TODOList'),
+                  content: const Text('Would you like to remove?'),
+                  textOK: const Text('Yes'),
+                  textCancel: const Text('No'),
+                )) {
+                  handleDeleteTask(item.id); // gọi hàm xoá
+                  return;
+                }
+                SizedBox(height: 12);
+                return print('pressedCancel');
               },
+
               child: Icon(
                 Icons.delete_outlined,
                 color: Color.fromARGB(255, 0, 0, 0),
